requirements.txt
numpy~=2.1.1
torch~=2.9.1
pillow~=12.0.0
tqdm~=4.66.5
torchvision~=0.24.1
iopath~=0.1.10
hydra-core~=1.3.2
omegaconf~=2.3.0
opencv-python~=4.12.0.88
supervision~=0.27.0
pycocotools~=2.0.10
transformers~=4.57.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth==2.35.0
requests==2.32.3
python-multipart==0.0.9